:truck: rename channel components
