+++ src/pages/channel/ui/channel/list-item.tsx
