+++ src/pages/channel/ui/channel/list.tsx
@@ -1,13 +1,13 @@
 'use client';
 
-import { useChannelStore } from '../store/channel';
+import ChannelListItem from './list-item';
+import { useChannelStore } from '../../store/channel';
 
 const ChannelList = () => {
   const channels = useChannelStore((state) => state.channels);
 
-  // TODO: replace with channel item component
   return channels.map((channel) => (
-    <div key={channel.link}>{channel.title}</div>
+    <ChannelListItem key={channel.link} channel={channel} />
   ));
 };
